+++ app/src/app/restaurant/[slug]/menu/page.tsx
@@ -1,25 +1,12 @@
+import NavBar from '@/app/components/NavBar';
 import Link from 'next/link';
 
 export default function RestaurantMenu() {
   return (
     <main className="bg-gray-100 min-h-screen w-screen">
       <main className="max-w-screen-2xl m-auto bg-white">
-        {/* NAVBAR */}
-        <nav className="bg-white p-2 flex justify-between">
-          <Link href="/" className="font-bold text-gray-700 text-2xl">
-            {' '}
-            OpenTable{' '}
-          </Link>
-          <div>
-            <div className="flex">
-              <button className="bg-blue-400 text-white border p-1 px-4 rounded mr-3">
-                Sign in
-              </button>
-              <button className="border p-1 px-4 rounded">Sign up</button>
-            </div>
-          </div>
-        </nav>
-        {/* NAVBAR */} {/* HEADER */}
+        <NavBar />
+        {/* HEADER */}
         <div className="h-96 overflow-hidden">
           <div className="bg-center bg-gradient-to-r from-[#0f1f47] to-[#5f6984] h-full flex justify-center items-center">
             <h1 className="text-7xl text-white captitalize text-shadow text-center">
